requirements.txt
# HomeOS Backend Dependencies
flask==3.0.0
flask-cors==4.0.0
python-telegram-bot==20.7